server/requirements.txt
opencv-python
numpy
pyzmq
pycocotools
torch
torchvision
pygame
pillow
argparse